Renamed occupations to tiles
