+++ src/components/table-container/TableCell.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import { hot } from "react-hot-loader";
 
 import { useAppSelector } from "../../redux/hooks";
-import * as OccupationsSlice from "../../redux/occupationsSlice";
+import * as TilesSlice from "../../redux/tilesSlice";
 
 import Tile from "./Tile";
 
@@ -31,7 +31,7 @@ function TableCell(props: Props): JSX.Element {
   return <div className="table-cell">{tile}</div>;
 }
 
-function useTileDataAt(x: string, y: number): OccupationsSlice.TileData | undefined {
+function useTileDataAt(x: string, y: number): TilesSlice.TileData | undefined {
   return useAppSelector(state => {
     const occupationsForRoom = state.occupations[y];
     return (occupationsForRoom === undefined) ? undefined : occupationsForRoom[x];
