+++ src/components/table-container/Tile.tsx
@@ -4,7 +4,7 @@ import { AnyAction } from "@reduxjs/toolkit";
 
 import { useAppDispatch } from "../../redux/hooks";
 import * as GrabbedTileSlice from "../../redux/grabbedTileSlice";
-import * as OccupationsSlice from "../../redux/occupationsSlice";
+import * as TilesSlice from "../../redux/tilesSlice";
 
 import "./Tile.css";
 
@@ -24,7 +24,7 @@ type Action = {
 type Props = {
   x: string,
   y: number,
-  tileData: OccupationsSlice.TileData
+  tileData: TilesSlice.TileData
 };
 
 function Tile(props: Props): JSX.Element {
